% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{apply}
\alias{apply}
\alias{apply,FileArray-method}
\title{Apply functions over file array margins (extended)}
\usage{
apply(X, MARGIN, FUN, ..., simplify = TRUE)

\S4method{apply}{FileArray}(X, MARGIN, FUN, ..., simplify = TRUE)
}
\arguments{
\item{X}{a file array}

\item{MARGIN}{scalar giving the subscripts which the function will be applied over. Current implementation only allows margin size to be one}

\item{FUN}{the function to be applied}

\item{...}{optional arguments to \code{FUN}}

\item{simplify}{a logical indicating whether results should be simplified if possible}
}
\value{
See Section 'Value' in \code{\link[base]{apply}};
}
\description{
Apply functions over file array margins (extended)
}
