% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{dim,FileRealizationSink-method}
\alias{dim,FileRealizationSink-method}
\title{Internal functions}
\usage{
\S4method{dim}{FileRealizationSink}(x)
}
\arguments{
\item{x}{FileArraySeed, FileArray, FileRealizationSink}
}
\description{
Internal functions that should not be used called directly.
}
