% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileArray-package.R
\docType{package}
\name{FileArray-package}
\alias{FileArray-package}
\alias{_PACKAGE}
\title{FileArray: Using filematrix as a DelayedArray Backend}
\description{
Implements a DelayedArray backend for reading and writing dense matrices in the filematrix format. Supports filematrix-backed SummarizedExperiment objects.
}
\keyword{internal}
