% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-docs.R
\name{filearray_threads}
\alias{filearray_threads}
\title{Set or get file array threads}
\usage{
filearray_threads(n, ...)
}
\arguments{
\item{n}{number of threads to set. If \code{n} is negative,
then default to the number of cores that computer has.}

\item{...}{internally used}
}
\value{
An integer of current number of threads
}
\description{
Will enable/disable multi-threaded reading or writing
at \code{C++} level.
}
