% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-typeof.R
\name{typeof}
\alias{typeof}
\alias{typeof,FileArray-method}
\alias{typeof,FileArrayProxy-method}
\title{The type of a file array (extended)}
\usage{
typeof(x)

\S4method{typeof}{FileArray}(x)

\S4method{typeof}{FileArrayProxy}(x)
}
\arguments{
\item{x}{any file array}
}
\value{
A character string. The possible values are \code{"double"},
\code{"integer"}, \code{"logical"}, and \code{"raw"}
}
\description{
The type of a file array (extended)
}
