% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-docs.R, R/methods-subset.R,
%   R/methods-subsetAssign.R, R/methods.R
\name{S3-filearray}
\alias{S3-filearray}
\alias{[.FileArray}
\alias{[<-.FileArray}
\alias{[[.FileArray}
\alias{as.array.FileArray}
\alias{dim.FileArray}
\alias{dimnames.FileArray}
\alias{dimnames<-.FileArray}
\alias{length.FileArray}
\alias{max.FileArray}
\alias{min.FileArray}
\alias{range.FileArray}
\alias{sum.FileArray}
\alias{subset.FileArray}
\title{'S3' methods for 'FileArray'}
\usage{
\method{[}{FileArray}(
  x,
  i,
  ...,
  drop = TRUE,
  reshape = NULL,
  strict = TRUE,
  dimnames = TRUE,
  split_dim = 0
)

\method{[}{FileArray}(x, i, ..., lazy = FALSE) <- value

\method{[[}{FileArray}(x, i)

\method{as.array}{FileArray}(x, reshape = NULL, drop = FALSE, ...)

\method{dim}{FileArray}(x)

\method{dimnames}{FileArray}(x)

\method{dimnames}{FileArray}(x) <- value

\method{length}{FileArray}(x)

\method{max}{FileArray}(x, na.rm = FALSE, ...)

\method{min}{FileArray}(x, na.rm = FALSE, ...)

\method{range}{FileArray}(x, na.rm = FALSE, ...)

\method{sum}{FileArray}(x, na.rm = FALSE, ...)

\method{subset}{FileArray}(x, ..., drop = FALSE, .env = parent.frame())
}
\arguments{
\item{x}{a file array}

\item{i, ...}{index set, or passed to other methods}

\item{drop}{whether to drop dimensions; see topic \code{\link[base]{Extract}}}

\item{reshape}{a new dimension to set before returning subset results; default is \code{NULL} (use default dimensions)}

\item{strict}{whether to allow indices to exceed bound; currently only accept \code{TRUE}}

\item{dimnames}{whether to preserve \code{\link[base]{dimnames}}}

\item{split_dim}{internally used; split dimension and calculate indices to
manually speed up the subset; value ranged from 0 to size of dimension minus
one.}

\item{lazy}{whether to lazy-evaluate the method, only works when assigning
arrays with logical array index}

\item{value}{value to substitute or set}

\item{na.rm}{whether to remove \code{NA} values during the calculation}

\item{.env}{environment to evaluate formula when evaluating subset margin indices.}
}
\description{
These are 'S3' methods for 'FileArray'
}
\section{Functions}{
\itemize{
\item \code{[}: subset array

\item \code{`[`(FileArray) <- value}: subset assign array

\item \code{[[}: get element by index

\item \code{as.array(FileArray)}: converts file array to native array in R

\item \code{dim(FileArray)}: get dimensions

\item \code{dimnames(FileArray)}: get dimension names

\item \code{dimnames(FileArray) <- value}: set dimension names

\item \code{length(FileArray)}: get array length

\item \code{max(FileArray)}: get max value

\item \code{min(FileArray)}: get min value

\item \code{range(FileArray)}: get value range

\item \code{sum(FileArray)}: get summation

\item \code{subset(FileArray)}: get subset file array with formulae

}}
