% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-docs.R, R/methods-add.R, R/methods-simple.R
\name{S4-filearray}
\alias{S4-filearray}
\alias{+,FileArray,FileArray-method}
\alias{+,FileArray,numeric-method}
\alias{+,numeric,FileArray-method}
\alias{+,FileArray,complex-method}
\alias{+,complex,FileArray-method}
\alias{+,FileArray,logical-method}
\alias{+,logical,FileArray-method}
\alias{+,FileArray,array-method}
\alias{+,array,FileArray-method}
\alias{-,FileArray,FileArray-method}
\alias{-,FileArray,numeric-method}
\alias{-,numeric,FileArray-method}
\alias{-,FileArray,complex-method}
\alias{-,complex,FileArray-method}
\alias{-,FileArray,logical-method}
\alias{-,logical,FileArray-method}
\alias{-,FileArray,array-method}
\alias{-,array,FileArray-method}
\alias{*,FileArray,FileArray-method}
\alias{*,FileArray,numeric-method}
\alias{*,numeric,FileArray-method}
\alias{*,FileArray,complex-method}
\alias{*,complex,FileArray-method}
\alias{*,FileArray,logical-method}
\alias{*,logical,FileArray-method}
\alias{*,FileArray,array-method}
\alias{*,array,FileArray-method}
\alias{/,FileArray,FileArray-method}
\alias{/,FileArray,numeric-method}
\alias{/,numeric,FileArray-method}
\alias{/,FileArray,complex-method}
\alias{/,complex,FileArray-method}
\alias{/,FileArray,logical-method}
\alias{/,logical,FileArray-method}
\alias{/,FileArray,array-method}
\alias{/,array,FileArray-method}
\alias{^,FileArray,FileArray-method}
\alias{^,FileArray,numeric-method}
\alias{^,numeric,FileArray-method}
\alias{^,FileArray,complex-method}
\alias{^,complex,FileArray-method}
\alias{^,FileArray,logical-method}
\alias{^,logical,FileArray-method}
\alias{^,FileArray,array-method}
\alias{^,array,FileArray-method}
\alias{\%\%,FileArray,FileArray-method}
\alias{\%\%,FileArray,numeric-method}
\alias{\%\%,numeric,FileArray-method}
\alias{\%\%,FileArray,complex-method}
\alias{\%\%,complex,FileArray-method}
\alias{\%\%,FileArray,logical-method}
\alias{\%\%,logical,FileArray-method}
\alias{\%\%,FileArray,array-method}
\alias{\%\%,array,FileArray-method}
\alias{\%/\%,FileArray,FileArray-method}
\alias{\%/\%,FileArray,numeric-method}
\alias{\%/\%,numeric,FileArray-method}
\alias{\%/\%,FileArray,complex-method}
\alias{\%/\%,complex,FileArray-method}
\alias{\%/\%,FileArray,logical-method}
\alias{\%/\%,logical,FileArray-method}
\alias{\%/\%,FileArray,array-method}
\alias{\%/\%,array,FileArray-method}
\alias{==,FileArray,FileArray-method}
\alias{==,FileArray,numeric-method}
\alias{==,numeric,FileArray-method}
\alias{==,FileArray,complex-method}
\alias{==,complex,FileArray-method}
\alias{==,FileArray,logical-method}
\alias{==,logical,FileArray-method}
\alias{==,FileArray,array-method}
\alias{==,array,FileArray-method}
\alias{>,FileArray,FileArray-method}
\alias{>,FileArray,numeric-method}
\alias{>,numeric,FileArray-method}
\alias{>,FileArray,complex-method}
\alias{>,complex,FileArray-method}
\alias{>,FileArray,logical-method}
\alias{>,logical,FileArray-method}
\alias{>,FileArray,array-method}
\alias{>,array,FileArray-method}
\alias{<,FileArray,FileArray-method}
\alias{<,FileArray,numeric-method}
\alias{<,numeric,FileArray-method}
\alias{<,FileArray,complex-method}
\alias{<,complex,FileArray-method}
\alias{<,FileArray,logical-method}
\alias{<,logical,FileArray-method}
\alias{<,FileArray,array-method}
\alias{<,array,FileArray-method}
\alias{!=,FileArray,FileArray-method}
\alias{!=,FileArray,numeric-method}
\alias{!=,numeric,FileArray-method}
\alias{!=,FileArray,complex-method}
\alias{!=,complex,FileArray-method}
\alias{!=,FileArray,logical-method}
\alias{!=,logical,FileArray-method}
\alias{!=,FileArray,array-method}
\alias{!=,array,FileArray-method}
\alias{>=,FileArray,FileArray-method}
\alias{>=,FileArray,numeric-method}
\alias{>=,numeric,FileArray-method}
\alias{>=,FileArray,complex-method}
\alias{>=,complex,FileArray-method}
\alias{>=,FileArray,logical-method}
\alias{>=,logical,FileArray-method}
\alias{>=,FileArray,array-method}
\alias{>=,array,FileArray-method}
\alias{<=,FileArray,FileArray-method}
\alias{<=,FileArray,numeric-method}
\alias{<=,numeric,FileArray-method}
\alias{<=,FileArray,complex-method}
\alias{<=,complex,FileArray-method}
\alias{<=,FileArray,logical-method}
\alias{<=,logical,FileArray-method}
\alias{<=,FileArray,array-method}
\alias{<=,array,FileArray-method}
\alias{&,FileArray,FileArray-method}
\alias{&,FileArray,numeric-method}
\alias{&,numeric,FileArray-method}
\alias{&,FileArray,complex-method}
\alias{&,complex,FileArray-method}
\alias{&,FileArray,logical-method}
\alias{&,logical,FileArray-method}
\alias{&,FileArray,array-method}
\alias{&,array,FileArray-method}
\alias{|,FileArray,FileArray-method}
\alias{|,FileArray,numeric-method}
\alias{|,numeric,FileArray-method}
\alias{|,FileArray,complex-method}
\alias{|,complex,FileArray-method}
\alias{|,FileArray,logical-method}
\alias{|,logical,FileArray-method}
\alias{|,FileArray,array-method}
\alias{|,array,FileArray-method}
\alias{!,FileArray-method}
\alias{exp,FileArray-method}
\alias{expm1,FileArray-method}
\alias{log,FileArray-method}
\alias{log10,FileArray-method}
\alias{log2,FileArray-method}
\alias{log1p,FileArray-method}
\alias{abs,FileArray-method}
\alias{sqrt,FileArray-method}
\alias{sign,FileArray-method}
\alias{signif,FileArray-method}
\alias{trunc,FileArray-method}
\alias{floor,FileArray-method}
\alias{ceiling,FileArray-method}
\alias{round,FileArray-method}
\alias{acos,FileArray-method}
\alias{acosh,FileArray-method}
\alias{asin,FileArray-method}
\alias{asinh,FileArray-method}
\alias{atan,FileArray-method}
\alias{atanh,FileArray-method}
\alias{cos,FileArray-method}
\alias{cosh,FileArray-method}
\alias{cospi,FileArray-method}
\alias{sin,FileArray-method}
\alias{sinh,FileArray-method}
\alias{sinpi,FileArray-method}
\alias{tan,FileArray-method}
\alias{tanh,FileArray-method}
\alias{tanpi,FileArray-method}
\alias{gamma,FileArray-method}
\alias{lgamma,FileArray-method}
\alias{digamma,FileArray-method}
\alias{trigamma,FileArray-method}
\alias{Arg,FileArray-method}
\alias{Conj,FileArray-method}
\alias{Im,FileArray-method}
\alias{Mod,FileArray-method}
\alias{Re,FileArray-method}
\alias{is.na,FileArray-method}
\title{'S4' methods for \code{FileArray}}
\usage{
\S4method{+}{FileArray,FileArray}(e1, e2)

\S4method{+}{FileArray,numeric}(e1, e2)

\S4method{+}{numeric,FileArray}(e1, e2)

\S4method{+}{FileArray,complex}(e1, e2)

\S4method{+}{complex,FileArray}(e1, e2)

\S4method{+}{FileArray,logical}(e1, e2)

\S4method{+}{logical,FileArray}(e1, e2)

\S4method{+}{FileArray,array}(e1, e2)

\S4method{+}{array,FileArray}(e1, e2)

\S4method{-}{FileArray,FileArray}(e1, e2)

\S4method{-}{FileArray,numeric}(e1, e2)

\S4method{-}{numeric,FileArray}(e1, e2)

\S4method{-}{FileArray,complex}(e1, e2)

\S4method{-}{complex,FileArray}(e1, e2)

\S4method{-}{FileArray,logical}(e1, e2)

\S4method{-}{logical,FileArray}(e1, e2)

\S4method{-}{FileArray,array}(e1, e2)

\S4method{-}{array,FileArray}(e1, e2)

\S4method{*}{FileArray,FileArray}(e1, e2)

\S4method{*}{FileArray,numeric}(e1, e2)

\S4method{*}{numeric,FileArray}(e1, e2)

\S4method{*}{FileArray,complex}(e1, e2)

\S4method{*}{complex,FileArray}(e1, e2)

\S4method{*}{FileArray,logical}(e1, e2)

\S4method{*}{logical,FileArray}(e1, e2)

\S4method{*}{FileArray,array}(e1, e2)

\S4method{*}{array,FileArray}(e1, e2)

\S4method{/}{FileArray,FileArray}(e1, e2)

\S4method{/}{FileArray,numeric}(e1, e2)

\S4method{/}{numeric,FileArray}(e1, e2)

\S4method{/}{FileArray,complex}(e1, e2)

\S4method{/}{complex,FileArray}(e1, e2)

\S4method{/}{FileArray,logical}(e1, e2)

\S4method{/}{logical,FileArray}(e1, e2)

\S4method{/}{FileArray,array}(e1, e2)

\S4method{/}{array,FileArray}(e1, e2)

\S4method{^}{FileArray,FileArray}(e1, e2)

\S4method{^}{FileArray,numeric}(e1, e2)

\S4method{^}{numeric,FileArray}(e1, e2)

\S4method{^}{FileArray,complex}(e1, e2)

\S4method{^}{complex,FileArray}(e1, e2)

\S4method{^}{FileArray,logical}(e1, e2)

\S4method{^}{logical,FileArray}(e1, e2)

\S4method{^}{FileArray,array}(e1, e2)

\S4method{^}{array,FileArray}(e1, e2)

\S4method{\%\%}{FileArray,FileArray}(e1, e2)

\S4method{\%\%}{FileArray,numeric}(e1, e2)

\S4method{\%\%}{numeric,FileArray}(e1, e2)

\S4method{\%\%}{FileArray,complex}(e1, e2)

\S4method{\%\%}{complex,FileArray}(e1, e2)

\S4method{\%\%}{FileArray,logical}(e1, e2)

\S4method{\%\%}{logical,FileArray}(e1, e2)

\S4method{\%\%}{FileArray,array}(e1, e2)

\S4method{\%\%}{array,FileArray}(e1, e2)

\S4method{\%/\%}{FileArray,FileArray}(e1, e2)

\S4method{\%/\%}{FileArray,numeric}(e1, e2)

\S4method{\%/\%}{numeric,FileArray}(e1, e2)

\S4method{\%/\%}{FileArray,complex}(e1, e2)

\S4method{\%/\%}{complex,FileArray}(e1, e2)

\S4method{\%/\%}{FileArray,logical}(e1, e2)

\S4method{\%/\%}{logical,FileArray}(e1, e2)

\S4method{\%/\%}{FileArray,array}(e1, e2)

\S4method{\%/\%}{array,FileArray}(e1, e2)

\S4method{==}{FileArray,FileArray}(e1, e2)

\S4method{==}{FileArray,numeric}(e1, e2)

\S4method{==}{numeric,FileArray}(e1, e2)

\S4method{==}{FileArray,complex}(e1, e2)

\S4method{==}{complex,FileArray}(e1, e2)

\S4method{==}{FileArray,logical}(e1, e2)

\S4method{==}{logical,FileArray}(e1, e2)

\S4method{==}{FileArray,array}(e1, e2)

\S4method{==}{array,FileArray}(e1, e2)

\S4method{>}{FileArray,FileArray}(e1, e2)

\S4method{>}{FileArray,numeric}(e1, e2)

\S4method{>}{numeric,FileArray}(e1, e2)

\S4method{>}{FileArray,complex}(e1, e2)

\S4method{>}{complex,FileArray}(e1, e2)

\S4method{>}{FileArray,logical}(e1, e2)

\S4method{>}{logical,FileArray}(e1, e2)

\S4method{>}{FileArray,array}(e1, e2)

\S4method{>}{array,FileArray}(e1, e2)

\S4method{<}{FileArray,FileArray}(e1, e2)

\S4method{<}{FileArray,numeric}(e1, e2)

\S4method{<}{numeric,FileArray}(e1, e2)

\S4method{<}{FileArray,complex}(e1, e2)

\S4method{<}{complex,FileArray}(e1, e2)

\S4method{<}{FileArray,logical}(e1, e2)

\S4method{<}{logical,FileArray}(e1, e2)

\S4method{<}{FileArray,array}(e1, e2)

\S4method{<}{array,FileArray}(e1, e2)

\S4method{!=}{FileArray,FileArray}(e1, e2)

\S4method{!=}{FileArray,numeric}(e1, e2)

\S4method{!=}{numeric,FileArray}(e1, e2)

\S4method{!=}{FileArray,complex}(e1, e2)

\S4method{!=}{complex,FileArray}(e1, e2)

\S4method{!=}{FileArray,logical}(e1, e2)

\S4method{!=}{logical,FileArray}(e1, e2)

\S4method{!=}{FileArray,array}(e1, e2)

\S4method{!=}{array,FileArray}(e1, e2)

\S4method{>=}{FileArray,FileArray}(e1, e2)

\S4method{>=}{FileArray,numeric}(e1, e2)

\S4method{>=}{numeric,FileArray}(e1, e2)

\S4method{>=}{FileArray,complex}(e1, e2)

\S4method{>=}{complex,FileArray}(e1, e2)

\S4method{>=}{FileArray,logical}(e1, e2)

\S4method{>=}{logical,FileArray}(e1, e2)

\S4method{>=}{FileArray,array}(e1, e2)

\S4method{>=}{array,FileArray}(e1, e2)

\S4method{<=}{FileArray,FileArray}(e1, e2)

\S4method{<=}{FileArray,numeric}(e1, e2)

\S4method{<=}{numeric,FileArray}(e1, e2)

\S4method{<=}{FileArray,complex}(e1, e2)

\S4method{<=}{complex,FileArray}(e1, e2)

\S4method{<=}{FileArray,logical}(e1, e2)

\S4method{<=}{logical,FileArray}(e1, e2)

\S4method{<=}{FileArray,array}(e1, e2)

\S4method{<=}{array,FileArray}(e1, e2)

\S4method{&}{FileArray,FileArray}(e1, e2)

\S4method{&}{FileArray,numeric}(e1, e2)

\S4method{&}{numeric,FileArray}(e1, e2)

\S4method{&}{FileArray,complex}(e1, e2)

\S4method{&}{complex,FileArray}(e1, e2)

\S4method{&}{FileArray,logical}(e1, e2)

\S4method{&}{logical,FileArray}(e1, e2)

\S4method{&}{FileArray,array}(e1, e2)

\S4method{&}{array,FileArray}(e1, e2)

\S4method{|}{FileArray,FileArray}(e1, e2)

\S4method{|}{FileArray,numeric}(e1, e2)

\S4method{|}{numeric,FileArray}(e1, e2)

\S4method{|}{FileArray,complex}(e1, e2)

\S4method{|}{complex,FileArray}(e1, e2)

\S4method{|}{FileArray,logical}(e1, e2)

\S4method{|}{logical,FileArray}(e1, e2)

\S4method{|}{FileArray,array}(e1, e2)

\S4method{|}{array,FileArray}(e1, e2)

\S4method{!}{FileArray}(x)

\S4method{exp}{FileArray}(x)

\S4method{expm1}{FileArray}(x)

\S4method{log}{FileArray}(x, base = exp(1))

\S4method{log10}{FileArray}(x)

\S4method{log2}{FileArray}(x)

\S4method{log1p}{FileArray}(x)

\S4method{abs}{FileArray}(x)

\S4method{sqrt}{FileArray}(x)

\S4method{sign}{FileArray}(x)

\S4method{signif}{FileArray}(x, digits = 6)

\S4method{trunc}{FileArray}(x, ...)

\S4method{floor}{FileArray}(x)

\S4method{ceiling}{FileArray}(x)

\S4method{round}{FileArray}(x, digits = 0)

\S4method{acos}{FileArray}(x)

\S4method{acosh}{FileArray}(x)

\S4method{asin}{FileArray}(x)

\S4method{asinh}{FileArray}(x)

\S4method{atan}{FileArray}(x)

\S4method{atanh}{FileArray}(x)

\S4method{cos}{FileArray}(x)

\S4method{cosh}{FileArray}(x)

\S4method{cospi}{FileArray}(x)

\S4method{sin}{FileArray}(x)

\S4method{sinh}{FileArray}(x)

\S4method{sinpi}{FileArray}(x)

\S4method{tan}{FileArray}(x)

\S4method{tanh}{FileArray}(x)

\S4method{tanpi}{FileArray}(x)

\S4method{gamma}{FileArray}(x)

\S4method{lgamma}{FileArray}(x)

\S4method{digamma}{FileArray}(x)

\S4method{trigamma}{FileArray}(x)

\S4method{Arg}{FileArray}(z)

\S4method{Conj}{FileArray}(z)

\S4method{Im}{FileArray}(z)

\S4method{Mod}{FileArray}(z)

\S4method{Re}{FileArray}(z)

\S4method{is.na}{FileArray}(x)
}
\arguments{
\item{x, z, e1, e2}{\code{FileArray} or compatible data}

\item{base, digits, ...}{passed to other methods}
}
\value{
See \code{\link[methods]{S4groupGeneric}}
}
\description{
'S4' methods for \code{FileArray}
}
